pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot_tests"
version = "1.0.0"
description = "Playwright automation tests for volto-chatbot"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.57.0",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "pytest-asyncio>=1.3.0",
    "pytest-html>=4.1.1",
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.1",
    "anyio>=4.12.1",
    "jinja2>=3.1.6",
    "litellm>=1.81.3",
    "numpy>=2.4.1",
    "sentence-transformers>=5.2.2",
    "md2pdf>=3.1.0"
]

[project.scripts]
chatbot_tests = "chatbot_tests.main:main"

[tool.setuptools]
# packages = ["chatbot_tests", "chatbot_tests.page_objects", "chatbot_tests.validators", "chatbot_tests.tests"]
packages = ["chatbot_tests", "chatbot_tests.page_objects"]

[tool.setuptools.package-dir]
chatbot_tests = "."
"chatbot_tests.page_objects" = "page_objects"
# "chatbot_tests.validators" = "validators"
# "chatbot_tests.tests" = "tests"

[tool.setuptools.package-data]
chatbot_tests = ["fixtures/*.json", ".env"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "basic: Basic chatbot functionality tests",
    "golden: Golden dataset validation tests",
    "halloumi: Halloumi fact-check tests",
    "feedback: Feedback functionality tests",
    "follow_up: Follow-up query tests",
    "slow: Marks tests as slow",
]
addopts = "-v --tb=short --strict-markers"
